Rename resultsmodifier front-end pages
