+++ app/manager-frontend/src/contexts/BackendApi.jsx
@@ -5,7 +5,7 @@ import { getDataSourceApi } from "api/datasource";
 import ky from "ky";
 import { createContext, useCallback, useContext, useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { getResultsModifierApi } from "api/resultsmodifier";
+import { getResultsModifierApi } from "api/resultsmodifiers";
 const BackendApiContext = createContext({});
 export const useBackendApi = () => useContext(BackendApiContext);
